Update model trainer & trainer config
