+++ netspresso/trainer/enums/__init__.py
@@ -1,5 +1,4 @@
-from .task import Task
-from .model import Backbone, Head, SUPPORTED_MODELS
+from .model import Task, Backbone, Head, SUPPORTED_MODELS
 from .data import Format
