+++ netspresso/trainer/enums/model.py
@@ -1,10 +1,12 @@
-from dataclasses import dataclass, field
-from pathlib import Path
-from typing import Any, Dict, List, Optional, Union
 from enum import Enum
 
 from netspresso_trainer.cfg.model import *
-from netspresso_trainer.cfg.model import ArchitectureConfig
+
+
+class Task(str, Enum):
+    IMAGE_CLASSIFICATION = "classification"
+    OBJECT_DETECTION = "detection"
+    SEMANTIC_SEGMENTATION = "segmentation"
 
 
 class Backbone(Enum):
@@ -15,7 +17,7 @@ class Backbone(Enum):
     ResNet = "resnet"
     SegFormer = "segformer"
     ViT = "vit"
-    YOLOX = "yolox"
+    CSPDarkNet = "cspdarknet"
 
 
 class Head(Enum):
@@ -24,31 +26,6 @@ class Head(Enum):
     FPN = "fpn"
     YOLOX_Head = "yolox_head"
     MLP_Decoder = "all_mlp_decoder"
-
-
-@dataclass
-class YOLOXArchitectureConfig(ArchitectureConfig):
-    backbone: Dict[str, Any] = field(default_factory=lambda: {
-        "name": "cspdarknet",
-        "params": {
-            "dep_mul": 0.33,
-            "wid_mul": 0.5,
-            "act_type": "silu",
-        },
-        "stage_params": None,
-    })
-
-
-@dataclass
-class DetectionYOLOXModelConfig(ModelConfig):
-    task: str = "detection"
-    checkpoint: Optional[Union[Path, str]] = "./weights/yolox/yolox_s.pth"
-    architecture: ArchitectureConfig = field(default_factory=lambda: YOLOXArchitectureConfig(
-        head={"name": "yolox_head"}
-    ))
-    losses: List[Dict[str, Any]] = field(default_factory=lambda: [
-        {"criterion": "yolox_loss", "weight": None},
-    ])
 
 
 SUPPORTED_MODELS = {
@@ -64,5 +41,5 @@ SUPPORTED_MODELS = {
     (Backbone.SegFormer, Head.FC): ClassificationSegFormerModelConfig(),
     (Backbone.SegFormer, Head.MLP_Decoder): SegmentationSegFormerModelConfig(),
     (Backbone.ViT, Head.FC): ClassificationViTModelConfig(),
-    (Backbone.YOLOX, Head.YOLOX_Head): DetectionYOLOXModelConfig(),
+    (Backbone.CSPDarkNet, Head.YOLOX_Head): DetectionYoloXModelConfig(),
 }
